# src/main/java/pl/iodkovskaya/travel_allowance_calculation/logic/overnightStay/model/dto/OvernightStayDto.java
package pl.iodkovskaya.travel_allowance_calculation.logic.overnightStay.model.dto;

import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotNull;
import lombok.AccessLevel;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;

@Getter
@AllArgsConstructor
@NoArgsConstructor(access = AccessLevel.PRIVATE)
public class OvernightStayDto {
    @NotNull(message = "Number of overnight stay without invoice cannot be null")
    @Min(value = 0, message = "Number of overnight stay without invoice cannot be negative")
    private Integer inputQuantityOfOvernightStayWithoutInvoice;

    @NotNull(message = "overnightStayWithoutInvoice cannot be null")
    private BigDecimal amountOfTotalOvernightsStayWithoutInvoice;

    @NotNull(message = "Number of overnight stay with invoice cannot be null")
    @Min(value = 0, message = "Number of overnight stay with invoice cannot be negative")
    private Integer inputQuantityOfOvernightStayWithInvoice;

    @NotNull(message = "overnightStayWithInvoice cannot be null")
    private BigDecimal amountOfTotalOvernightsStayWithInvoice;

    private boolean isAllowedMoreHigherPayment;
}